Rename Swatch component to SwatchEditor
