+++ components/SwatchEditor.jsx
